+++ src/subcommands/profile/create.rs
@@ -1,5 +1,5 @@
 use anyhow::{bail, Result};
-use dialoguer::{theme::ColorfulTheme, Confirm, Input, Select};
+use dialoguer::{Confirm, Input, Select};
 use ferinth::Ferinth;
 use libium::{config, file_picker, misc};
 use std::path::PathBuf;
@@ -9,7 +9,7 @@ pub async fn create(
     config: &mut config::structs::Config,
     game_version: Option<String>,
     force_game_version: bool,
-    mod_loader: Option<config::structs::ModLoaders>,
+    mod_loader: Option<config::structs::ModLoader>,
     name: Option<String>,
     output_dir: Option<PathBuf>,
 ) -> Result<()> {
@@ -53,7 +53,7 @@ pub async fn create(
             // Let user pick mods directory
             let mut selected_mods_dir = misc::get_mods_dir();
             println!("The default mods directory is {:?}", selected_mods_dir);
-            if Confirm::with_theme(&ColorfulTheme::default())
+            if Confirm::with_theme(&*crate::THEME)
                 .with_prompt("Would you like to specify a custom mods directory?")
                 .interact()?
             {
@@ -65,7 +65,7 @@ pub async fn create(
             let mut name = String::new();
             let mut prompt = true;
             while prompt {
-                name = Input::with_theme(&ColorfulTheme::default())
+                name = Input::with_theme(&*crate::THEME)
                     .with_prompt("What should this profile be called?")
                     .interact_text()?;
 
@@ -82,31 +82,19 @@ pub async fn create(
             // Let user pick Minecraft version
             let mut latest_versions: Vec<String> = misc::get_major_mc_versions(10).await?;
             println!();
-            let selected_version = Select::with_theme(&ColorfulTheme::default())
+            let selected_version = Select::with_theme(&*crate::THEME)
                 .with_prompt("Which version of Minecraft do you play?")
                 .items(&latest_versions)
                 .default(0)
                 .interact()?;
             let selected_version = latest_versions.swap_remove(selected_version);
 
-            // Let user pick mod loader
-            let mod_loaders = ["Fabric", "Forge"];
-            let selected_loader = if Select::with_theme(&ColorfulTheme::default())
-                .with_prompt("Which mod loader do you use?")
-                .items(&mod_loaders)
-                .interact()?
-                == 0
-            {
-                config::structs::ModLoaders::Fabric
-            } else {
-                config::structs::ModLoaders::Forge
-            };
             config.profiles.push(config::structs::Profile {
                 name,
                 output_dir: selected_mods_dir,
                 mods: Vec::new(),
                 game_version: selected_version,
-                mod_loader: selected_loader,
+                mod_loader: super::pick_mod_loader(None)?,
             });
         },
         _ => {