+++ src/subcommands/profile/delete.rs
@@ -1,6 +1,6 @@
 use crate::subcommands::switch::switch;
 use anyhow::{bail, Result};
-use dialoguer::{theme::ColorfulTheme, Select};
+use dialoguer::Select;
 use libium::config;
 
 pub fn delete(config: &mut config::structs::Config, profile_name: Option<String>) -> Result<()> {
@@ -23,7 +23,7 @@ pub fn delete(config: &mut config::structs::Config, profile_name: Option<String>
                 .map(|profile| &profile.name)
                 .collect::<Vec<_>>();
 
-            let selection = Select::with_theme(&ColorfulTheme::default())
+            let selection = Select::with_theme(&*crate::THEME)
                 .with_prompt("Select which profile to delete")
                 .items(&profile_names)
                 .default(config.active_profile)